Fixed memory corruption due to strcat() function
- The problem with memory stack corruption occurred in the 'fetch_saveimagepath()' function when starting the emulator.
Because the memory allocated on the stack `TCHAR path[MAX_DPATH]` in function `DISK_get_default_saveimagepath` does not have a end of line marker ('/0').
That's why further call of the strcat() function which first searches for a zero byte and then adds another string to it resulted in stack corruption in this place and application crash as a consequence.
- In other similar functions, the same scenario is possible.